Basic top tabs working on iOS, still lots of behavior to fix
